Add test-shell, fix /bin/true
Fix /bin/true by changing it to an empty executable file
for all platforms.  The reason for this is that true is a symlink
to busybox and there are other programs symlinked to true.  When
one of these second order symlinks is executed, busybox with see its
name and _not_ "true" do it will issue an applet not found message;
it will be even more confusing if busybox recognises the name!